+++ ui/src/components/mail/mail-list.tsx
@@ -189,7 +189,7 @@ export function MailList() {
         >
           {allMails.map((mail, index) =>
             isUnifiedView ? (
-              <DragMail key={mail.id} mailId={mail.id} accountId={mail.account_id}>
+              <DragMail key={mail.id} mailId={mail.id} accountId={mail.account_id} folderId={mail.folder_id}>
                 <UnifiedMailItem
                   mail={mail as UnifiedMessageSummary}
                   isSelected={mail.id === selectedMailId}
@@ -202,7 +202,7 @@ export function MailList() {
                 />
               </DragMail>
             ) : (
-              <DragMail key={mail.id} mailId={mail.id} accountId={mail.account_id}>
+              <DragMail key={mail.id} mailId={mail.id} accountId={mail.account_id} folderId={mail.folder_id}>
                 <MailListItem
                   mail={mail as MessageSummary}
                   isSelected={mail.id === selectedMailId}